% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_urls.R
\name{buildDataRequesttUrl}
\alias{buildDataRequesttUrl}
\title{Create a url to request data for a site and measurement from a Hilltop
Server.}
\usage{
buildDataRequesttUrl(endpoint, site, measurement, from, to, tsType,
  alignment, method, interval)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}

\item{site}{string A site that has data available on the server.}

\item{measurement}{string A measurement that has data available for the
requested site on the server.}

\item{from}{string (optional) The date (or date time combination) that data
is required from (if not provided only the last measurement will be
provided.)}

\item{to}{string (optional) The date (or date time combination) that data is
required up to (if not provided data will be provided from the from date to
the most recent data available.)}

\item{tsType}{string (optional) The type of time series being requested.
Default is StdTimeSeries, but quality code data can be obtained using
tsType = StdQualSeries.}

\item{alignment}{string (optional) A time or keyword that the data should be
aligned to.}

\item{method}{string (optional) A statistical method to apply to the data.
The options are Interpolate, Average, Total, Moving Average and EP.
Interpolate - Provide data aligned with the requested interval. Average -
Provide the average of the measurement across the interval. Total - Provide
the total in the interval. Moving Average - Provide the moving average in
the interval, timestamped to half of the interval. EP - Provide the
exceedence percentile based on the whole available dataset.}

\item{interval}{string (optional) The interval for the statistic to be
computed over.}
}
\value{
string A Hilltop url request for data for the site and measurement in
  the time period for the site.
}
\description{
Takes a valid Hilltop server endpoint, a site name, a measurement name  and
optional from and to arguments and an optional timeseries type argument and
returns a url that requests the data for that site and measurement for the
requested time range from the endpoint. For standard time series, statistics
can be requested, to request these a method, interval and allignment argument
need to be provided.
}
