% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_urls.R
\name{buildEnsembleStatsRequestUrl}
\alias{buildEnsembleStatsRequestUrl}
\title{Create a url to request ensemble statistics for a site and measurement.}
\usage{
buildEnsembleStatsRequestUrl(
  endpoint,
  site,
  measurement,
  from = NULL,
  to = NULL,
  timeInterval = NULL,
  statistic = "MonthlyPDF",
  lowerPercentile = NULL,
  upperPercentile = NULL
)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}

\item{site}{string A site that has data available on the server.}

\item{measurement}{string A measurement that has data available for the
requested site on the server.}

\item{from}{string (optional) The date (or date time combination) that data
is required from (if not provided only the last measurement will be
provided.)}

\item{to}{string (optional) The date (or date time combination) that data is
required up to (if not provided data will be provided from the from date to
the most recent data available.)}

\item{timeInterval}{string (optional) A time interval that data is requested
over, either of the form P1D, or P3M, or fromDateTime/toDateTime.}

\item{statistic}{string The statistic function to use.  Options are
HourlyExtrema, DailyExtrema, MonthlyExtrema,
MeanHourlyExtrema, MeanDailyExtrema, MeanMonthlyExtrema,
ExtremeHourlyMean, ExtremeDailyMean, ExtremeMonthlyMean,
HourlyPDF, DailyPDF, MonthlyPDF}

\item{lowerPercentile}{string (optional) The lower percentile value to return from a PDF request.}

\item{upperPercentile}{string (optional) The upper percentile value to return from a PDF request.}
}
\value{
string A Hilltop url request for data for the site and measurement in
  the time period for the site.
}
\description{
\code{buildEnsembleStatsRequestUrl} returns a url to request ensemble statisticss for a site and measurement
}
\details{
Takes a valid Hilltop server endpoint, a site name, a measurement name  and
optional from, to or timeIntervaland arguments, and a Statistic
argument and returns a url that requests the ensemble statistics for that site and
measurement for the requested time range from the endpoint.
}
