% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_urls.R
\name{buildMeasurementListUrl}
\alias{buildMeasurementListUrl}
\title{Create a url to request a measurement list for a site from a Hilltop Server.}
\usage{
buildMeasurementListUrl(endpoint, site)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}

\item{site}{string A site that has data available on the server.}
}
\value{
string A Hilltop url request for a measurement list for the site.
}
\description{
Takes a valid Hilltop server endpoint and a site name and returns a url that
requests the measurement list for that site from the endpoint.
}
