% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_urls.R
\name{buildSiteListUrl}
\alias{buildSiteListUrl}
\title{Create a url to request a site list from a Hilltop Server.}
\usage{
buildSiteListUrl(endpoint, location)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}

\item{location}{logical (optional) whether the Lat Long for the sites is
returned}
}
\value{
string A Hilltop url request for a site list.
}
\description{
\code{buildSiteListUrl} returns a url to request a site list
}
\details{
Takes a valid Hilltop server endpoint and returns a url that requests the
site list from the endpoint.
}
