% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_endpointChecks.R
\name{checkFixEndpoint}
\alias{checkFixEndpoint}
\title{Checks and fixes a Hiltop endpoint.}
\usage{
checkFixEndpoint(endpoint)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}
}
\value{
A valid Hilltop endpoint.
}
\description{
Helper function to check and fix a Hilltop endpoint string.
}
