% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_requests.R
\name{getHilltopEnsembleStats}
\alias{getHilltopEnsembleStats}
\title{General function to get ensemblestats data from a Hilltop server.}
\usage{
getHilltopEnsembleStats(
  endpoint,
  site,
  measurement,
  from = NULL,
  to = NULL,
  timeInterval = NULL,
  statistic = "MonthlyPDF",
  lowerPercentile = NULL,
  upperPercentile = NULL,
  byYear = FALSE
)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}

\item{site}{string A site that has data available on the server.}

\item{measurement}{string A measurement that has data available for the
requested site on the server.}

\item{from}{string (optional) The date (or date time combination) that data
is required from (if not provided only the last measurement will be
provided.)}

\item{to}{string (optional) The date (or date time combination) that data is
required up to (if not provided data will be provided from the from date to
the most recent data available.)}

\item{timeInterval}{string (optional) A time interval that data is requested
over, either of the form P1D, or P3M, or fromDateTime/toDateTime.}

\item{statistic}{string The statistic function to use.  Options are
HourlyExtrema, DailyExtrema, MonthlyExtrema,
MeanHourlyExtrema, MeanDailyExtrema, MeanMonthlyExtrema,
ExtremeHourlyMean, ExtremeDailyMean, ExtremeMonthlyMean,
HourlyPDF, DailyPDF, MonthlyPDF}

\item{lowerPercentile}{string (optional) The lower percentile value to return from a PDF request.}

\item{upperPercentile}{string (optional) The upper percentile value to return from a PDF request.}

\item{byYear}{boolean A flag to indicate whether you want ensemble stats for each year.
Note: If true then the returned data will be for full years
based on the from year and to year where data is available}
}
\value{
dataframe A dataframe of the ensemble stats.
}
\description{
\code{getHilltopEnsembleStats} retrieves time series data from a Hilltop server.
}
\details{
Takes a valid Hilltop server endpoint, a site name, a measurement name  and
optional from, to or timeIntervaland arguments, a Statistic and
optional upper and lower percentile arguments and byYear argument.
Returns a dataframe of the relevant Ensemble Statistics for that site and measurement
for the requested time range from the endpoint.
}
