% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_requests.R
\name{getHilltopMeasurements}
\alias{getHilltopMeasurements}
\title{General funtion to get a measuremnt list for a site from a Hilltop server.}
\usage{
getHilltopMeasurements(endpoint, site)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}

\item{site}{string A site that has data available on the server.}
}
\value{
A dataframe of measurements and associated metadata
}
\description{
Takes a valid hilltop endpoint and site, and returns a list of available measurements
available from the endpoint.
}
