% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_requests.R
\name{getHilltopSites}
\alias{getHilltopSites}
\title{General funtion to get a site list from a Hilltop server.}
\usage{
getHilltopSites(endpoint)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}
}
\value{
A dataframe of site names
}
\description{
\code{getHilltopSites} returns a list of sites.
}
\details{
Takes a valid hilltop endpoint, and returns a list of available sites
available from the endpoint.
}
