% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hill_func}
\alias{hill_func}
\title{\code{hill_func} to calculate functional diversity for each site.}
\usage{
hill_func(comm, traits, traits_as_is = FALSE, q = 0, base = exp(1),
  checkdata = TRUE, ...)
}
\arguments{
\item{comm}{data frame of vegtation data. Sites as rows, species as columns.}

\item{traits}{data frame of species functional traits data. Species as rows, traits as columns.
It can include both continuous and categorical data. It will be transformed into a distance
matrix using `FD::gowdis(traits)`.}

\item{traits_as_is}{if FALSE (default) traits data frame will be transformed into a distance}

\item{q}{hill number, q (default is 0) to control weights of species abundance.}

\item{base}{default is exp(1), the base of log.}

\item{checkdata}{default is TRUE.}
}
\value{
a matrix, with these information for each site: Q (Rao's Q); D_q (functional hill number,
 the effective number of equally abundant and functionally equally distince species);
 MD_q (mean functional diversity per species, the effective sum of pairwise distances between
 a fixed species and all other species); FD_q (total functional diversity, the effective total functional
 distance between species of the assemblage). See Chiu and Chao 2014 page 4 for more information.
}
\description{
\code{hill_func} to calculate functional diversity for each site.
}
\examples{
library(FD); data(dummy)
hill_func(comm = dummy$abun, traits = dummy$trait, q = 0)
hill_func(comm = dummy$abun, traits = dummy$trait, q = 1)
hill_func(comm = dummy$abun, traits = dummy$trait, q = 0.9999)
hill_func(comm = dummy$abun, traits = dummy$trait, q = 2)
hill_func(comm = dummy$abun, traits = dummy$trait, q = 3)
}
\author{
Daijiang Li
}

