% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_phylo.R, R/hill_phylo_parti.R,
%   R/hill_phylo_parti_pairwise.R
\name{hill_phylo}
\alias{hill_phylo}
\alias{hill_phylo_parti}
\alias{hill_phylo_parti_pairwise}
\title{Phylogenetic diversity through Hill Numbers}
\usage{
hill_phylo(comm, tree, q = 0, base = exp(1), rel_then_pool = TRUE,
  show.warning = TRUE)

hill_phylo_parti(comm, tree, q = 0, base = exp(1), rel_then_pool = TRUE,
  show.warning = TRUE)

hill_phylo_parti_pairwise(comm, tree, q = 0, output = c("data.frame",
  "matrix"), pairs = c("unique", "full"), ...)
}
\arguments{
\item{comm}{data frame of vegtation data. Sites as rows, species as columns.}

\item{tree}{a phylogeny with class "phylo".}

\item{q}{hill number, any non-negative value.}

\item{base}{default is exp(1), the base of log.}

\item{rel_then_pool}{default is TRUE. Abundance of species are changed to relative abundance within sites.}

\item{show.warning}{whether to print warning, default is TRUE.}

\item{...}{additional arguments for \code{hill_func_parti}.}
}
\value{
a vector of hill number based phylogenetic diversity for all sites.

a data frame with one row and six columns: q, gamma diversity, alpha diveristy,
beta diversity, local similarity, and region similarity.

a data frame or a matrix with results for all pairwise comparisons.
}
\description{
Phylogenetic diversity through Hill Numbers
}
\examples{
comm = dummy = FD::dummy$abun
tree = ape::rtree(n = ncol(comm), tip.label = paste0("sp", 1:8))
hill_phylo(comm, tree, q = 0)
hill_phylo(comm, tree, q = 0.999)
hill_phylo(comm, tree, q = 1)
hill_phylo(comm, tree, q = 2)
hill_phylo_parti(comm, tree, q = 0)
hill_phylo_parti(comm, tree, q = 0.999)
hill_phylo_parti(comm, tree, q = 1)
hill_phylo_parti(comm, tree, q = 2)

hill_phylo_parti_pairwise(comm, tree, q = 0, show.warning = FALSE)
hill_phylo_parti_pairwise(comm, tree, q = 0.999, show.warning = FALSE)
hill_phylo_parti_pairwise(comm, tree, q = 1, show.warning = FALSE)
hill_phylo_parti_pairwise(comm, tree, q = 2, show.warning = FALSE)

}
\seealso{
\code{\link{hill_phylo_parti}}
}
