% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_phylo_parti.R
\name{hill_phylo_parti}
\alias{hill_phylo_parti}
\title{Phylogenetic diversity of multiple sites}
\usage{
hill_phylo_parti(comm, tree, q = 0, base = exp(1),
  rel_then_pool = TRUE, show.warning = TRUE)
}
\arguments{
\item{comm}{data frame of vegtation data. Sites as rows, species as columns.}

\item{tree}{a phylogeny with class 'phylo'.}

\item{q}{hill number, any non-negative value.}

\item{base}{default is exp(1), the base of log.}

\item{rel_then_pool}{default is TRUE. Abundance of species are changed to relative abundance within sites.}

\item{show.warning}{whether to print warning, default is TRUE.}
}
\value{
a data frame with one row (across all sites) and six columns: q, gamma diversity, alpha diveristy,
beta diversity, local similarity, and region similarity.
}
\description{
Calculate overall phylogenetic diversity and site similarity across multiple sites.
}
\examples{
comm = dummy = FD::dummy$abun
tree = ape::rtree(n = ncol(comm), tip.label = paste0('sp', 1:8))
hill_phylo_parti(comm, tree, q = 0)
hill_phylo_parti(comm, tree, q = 0.999)
hill_phylo_parti(comm, tree, q = 1)
hill_phylo_parti(comm, tree, q = 2)

}
\references{
Chao, Anne, Chun-Huo Chiu, and Lou Jost. Unifying Species Diversity, Phylogenetic Diversity, Functional Diversity, and Related Similarity and Differentiation Measures Through Hill Numbers. Annual Review of Ecology, Evolution, and Systematics 45, no. 1 (2014): 297–324. <doi:10.1146/annurev-ecolsys-120213-091540>.
}
\author{
Chiu & Chao, Daijiang Li
}
