% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_taxa.R
\name{hill_taxa}
\alias{hill_taxa}
\title{Taxonomic diversity through Hill Numbers}
\usage{
hill_taxa(comm, q = 0, MARGIN = 1, base = exp(1))
}
\arguments{
\item{comm}{data frame of vegtation data. Sites as rows, species as columns.}

\item{q}{hill number, q = 0 (default) to get species richness,
q = 1 to get shannon entropy, q = 2 will give inverse Simpson.}

\item{MARGIN}{default is 1, if sites are columns, set MARGIN to 2.}

\item{base}{default is exp(1), the base of log.}
}
\value{
a named vector, diversity values for each site in the comm.
}
\description{
Calculate taxonomic diversity for each site (alpha diversity).
}
\examples{
dummy = FD::dummy
hill_taxa(comm = dummy$abun, q = 0)
# same as: vegan::specnumber(dummy$abun)
hill_taxa(comm = dummy$abun, q = 1)
# same as: exp(vegan::diversity(x = dummy$abun, index = "shannon"))
hill_taxa(comm = dummy$abun, q = 2)
# same as: vegan::diversity(x = dummy$abun, index = "invsimpson")
hill_taxa(comm = dummy$abun, q = 0.999)


}
\author{
Daijiang Li
}
