% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_taxa_parti.R
\name{hill_taxa_parti}
\alias{hill_taxa_parti}
\title{Decompostion of Taxonomic diversity through Hill Numbers}
\usage{
hill_taxa_parti(comm, q = 0, base = exp(1), rel_then_pool = TRUE)
}
\arguments{
\item{comm}{data frame of vegtation data. Sites as rows, species as columns.}

\item{q}{hill number, q = 0 (default) to get species richness, q = 1 to get shannon entropy, q = 2 will give inverse Simpson.}

\item{base}{default is exp(1), the base of log.}

\item{rel_then_pool}{default is TRUE. Abundance of species are first changed to relative abundance within sites,
then pooled into one assemblage. If FALSE, sites are pooled first, then change abundance of species
to relative abundance.}
}
\value{
a data frame with one row, including these columns: q, gamma diversity, alpha diveristy,
beta diversity, MacArthur's homogeneity measure, local species overlap, and region species overlap.
See Chao, Chiu and Jost 2014 Table 2 for more information.
}
\description{
Calculate taxonomic gamma, alpha, and beta diversity for communities, as
well as site similarity. If comm has 2 sites, this function gives pair comparison.
If comm has >2 sites, gamma diversity is the diversity of the pooled assemblage,
alpha is the average diversity across all site, beta is across all communities.
}
\examples{
dummy = FD::dummy
hill_taxa_parti(comm = dummy$abun, q = 0)
hill_taxa_parti(comm = dummy$abun, q = 1)
hill_taxa_parti(comm = dummy$abun, q = 0.9999999)
hill_taxa_parti(comm = dummy$abun, q = 0.9999999, rel_then_pool = FALSE)
hill_taxa_parti(comm = dummy$abun, q = 1, rel_then_pool = FALSE)
hill_taxa_parti(comm = dummy$abun, q = 2)
hill_taxa_parti(comm = dummy$abun, q = 3)


}
\seealso{
\code{\link{hill_taxa_parti}}
}
\author{
Daijiang Li
}
