% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{hillR}
\alias{hillR}
\alias{hillR-package}
\title{hillR: Diversity Through Hill Numbers}
\description{
The R package `hillR` implements the framework proposed by @chao2014unifying and makes it easy to calculate taxonomic, functional, and phylogenetic diversity of ecological communities as Hill numbers. For each facet of diversity, `hillR` has three functions. The first set of functions (\code{hill_taxa}, \code{hill_func}, and \code{hill_phylo}) calculates alpha diversity of each site. The second set of functions (\code{hill_taxa_parti}, \code{hill_func_parti}, and \code{hill_phylo_parti}) calculates diversity across all sites. The third set of functions (\code{hill_taxa_parti_pairwise}, \code{hill_func_parti_pairwise}, and \code{hill_phylo_parti_pairwise}) calculates all possible pairwise diversity across all sites. Users can set the argument _q_ to control the weight of species abundance.
}
\details{
Users may be interested in other similar packages such as \href{https://cran.r-project.org/web/packages/vegetarian/index.html}{vegetarian} and \href{https://cran.r-project.org/web/packages/iNEXT/index.html}{iNEXT}.
}
\section{Taxonomic Hill Numbers}{

\code{\link{hill_taxa}}, \code{\link{hill_taxa_parti}}, \code{\link{hill_taxa_parti_pairwise}}
}

\section{Functional Hill Numbers}{

\code{\link{hill_func}}, \code{\link{hill_func_parti}}, \code{\link{hill_func_parti_pairwise}}
}

\section{Phylogenetic Hill Numbers}{

\code{\link{hill_phylo}}, \code{\link{hill_phylo_parti}}, \code{\link{hill_phylo_parti_pairwise}}
}

