% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{hilltopAttributeHelper}
\alias{hilltopAttributeHelper}
\title{Hilltop XML Attribute Identifier.}
\usage{
hilltopAttributeHelper(xmlNode)
}
\arguments{
\item{xmlNode}{The XML from a Hilltop XML node.}
}
\value{
The name of the attribute described by the XML node.
}
\description{
Helper function to return the appropriate xml attribute name depending whether
the attribute of interest is from a named node, or is a named parameter.
}
