% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_measurements.R
\name{hilltopDataSourceToDF}
\alias{hilltopDataSourceToDF}
\title{Hilltop Data Source To Data Frame.}
\usage{
hilltopDataSourceToDF(dataxml)
}
\description{
Helper function that reads the nodes within a the DataSource ItemInfo node of a
Hilltop XML response from a anyXmlParse(url) request such as
dataxml<-anyXmlParse(url). Returns a dataframe of the Info for each Item.
Handles missing results and doen't require prior knowledge of the items.
}
