% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_measurements.R
\name{hilltopDepthProfile}
\alias{hilltopDepthProfile}
\title{Hilltop Depth Profile Measurement Request As A Data Frame.}
\usage{
hilltopDepthProfile(url)
}
\arguments{
\item{url}{A url that returns an XML document of a Hilltop Depth Profile.}
}
\value{
A dataframe of a depth profile from a Hilltop datasource.
}
\description{
\code{hilltopDepthProfile} returns a depth profile at times.
}
\details{
Function to process the XML from a Hilltop url request for Depth Profile data
into a dataframe that contains the measurement information. It returns a
dataframe of the data for depth and each profile timestamp, including
measurement name and the Site Name.
}
