% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_ensemble_stats.R
\name{hilltopEnsembleStatByTimePeriod}
\alias{hilltopEnsembleStatByTimePeriod}
\title{Hilltop Ensemble Statistics By Time Period.}
\usage{
hilltopEnsembleStatByTimePeriod(dataxml)
}
\description{
Helper function that takes parsed xml from an EnsembleStats Request. Returns
the statistics for each time period (depending whether hourly, monthly or
annual stats).
}
