% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_lists.R
\name{hilltopMeasInfoList}
\alias{hilltopMeasInfoList}
\title{Hilltop Data Sources And Measurement Information At A Site.}
\usage{
hilltopMeasInfoList(measlistxml)

hilltopMeasInfoList(measlistxml)
}
\description{
Helper function.
Takes an xml document from a Hilltop MeasurementList at a Site request.
Returns a dataframe of the measurement information and datasources.

Helper function.
Takes an xml document from a Hilltop MeasurementList at a Site request.
Returns a dataframe of the measurement information and datasources.
}
