% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_measurements.R
\name{hilltopMeasurementToDF}
\alias{hilltopMeasurementToDF}
\title{Hilltop Measurement To Data Frame.}
\usage{
hilltopMeasurementToDF(xmldata)
}
\arguments{
\item{xmldata}{An XML document, returned from a url request using anyXmlParse.}
}
\value{
A dataframe of data for each timestep.
}
\description{
Helper function that reads the nodes within a the Measurement node of a Hilltop
XML response from a anyXmlParse(url) request such as xmldata<-anyXmlParse(url).
Returns a dataframe of the data for each timestamp. Handles missing results and
doesn't require prior knowledge of parameter names. Handles true measurements
and WQ Sample requests.
}
