% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_lists.R
\name{hilltopSiteList}
\alias{hilltopSiteList}
\title{Get A List of Sites Available From A Hilltop Web Service.}
\usage{
hilltopSiteList(sitexml)

hilltopSiteList(sitexml)
}
\description{
Takes an parsed xml document from a Hilltop SiteList request.
Returns a dataframe of the available sites and location if available.

Takes an parsed xml document from a Hilltop SiteList request.
Returns a dataframe of the available sites and location if available.
}
