% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{hilltopValueHelper}
\alias{hilltopValueHelper}
\title{Hilltop XML Value Identifier.}
\usage{
hilltopValueHelper(xmlNode)
}
\arguments{
\item{xmlNode}{The XML from a Hilltop XML node.}
}
\value{
The value of the attribute described by the XML node.
}
\description{
\code{hilltopValueHelper} returns the value of parameters or nodes.
}
\details{
Helper function to return the appropriate xml value depending whether the value
of interest is from a named node, or is a named parameter.
}
