% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{hillXmlDataType}
\alias{hillXmlDataType}
\title{Checks what data type the returned XML is}
\usage{
hillXmlDataType(xmldata)
}
\arguments{
\item{xmldata}{An XML document, returned from a url request using anyXmlParse.}
}
\value{
character The data type of the Hilltop data,
options include SimpleTimeSeries, DepthProfile, GaugingResults.
}
\description{
\code{hillXmlDataType} returns the data type of the XML
}
