% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{hillXmlParse}
\alias{hillXmlParse}
\title{Parse hilltop XML, notifying errors}
\usage{
hillXmlParse(url)
}
\arguments{
\item{url}{A url that returns an XML document.}
}
\value{
Parsed Hilltop XML.
}
\description{
\code{hillXmlParse} returns hilltop xml for processing
}
