% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{is.hilltopXml}
\alias{is.hilltopXml}
\title{Checks Whether XML Is Hilltop XML.}
\usage{
is.hilltopXml(xmldata)
}
\arguments{
\item{xmldata}{An XML document, returned from a url request using anyXmlParse.}
}
\value{
logical TRUE / FALSE
}
\description{
\code{is.hilltopXml} Checks if an xml document is hilltop xml, returns True or False accordingly.
}
