% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{stringEnd}
\alias{stringEnd}
\title{Return the last n charachters from a string.}
\usage{
stringEnd(string, n)
}
\arguments{
\item{string}{string The string that the substring is to be extracted from.}

\item{n}{integer The number of charachters from the end of the string to return.}
}
\value{
string The last n charachters from the input string.
}
\description{
\code{stringEnd} return the last n charachters of a string.
}
