% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{xmlError}
\alias{xmlError}
\title{Return any error message provided by the server.}
\usage{
xmlError(xmldata)
}
\arguments{
\item{xmldata}{An XML document, returned from a url request using anyXmlParse.}
}
\value{
character The error message provided by the Hilltop server.
}
\description{
\code{xmlError} returns the error message from the server
}
