% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{anyXmlParse}
\alias{anyXmlParse}
\title{Parser For Hilltop XML}
\usage{
anyXmlParse(url)
}
\description{
Helper function to parse data from a hilltop server. Takes a valid url as an
input and returns a parsed xml document ready for other functions. Handles
https requests as well as http.
}
