% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_urls.R
\name{buildSiteListUrl}
\alias{buildSiteListUrl}
\title{Create a url to request a site list from a Hilltop Server.}
\usage{
buildSiteListUrl(endpoint)
}
\arguments{
\item{endpoint}{A valid hilltop endpoint.}
}
\value{
A Hilltop url request.
}
\description{
Takes a valid Hilltop server endpoint and returns a url that requests the
site list from the endpoint.
}
