% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_full_request.R
\name{fullGetHilltopData}
\alias{fullGetHilltopData}
\title{Multiple sites and measurements to Dataframe.}
\usage{
fullGetHilltopData(endpoint, sites, measurements, from, to, option,
  timeInterval, alignment, method, interval)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}

\item{sites}{string vector The sites that data is to be requested from.}

\item{measurements}{string vector The measurements that are to be requested.}

\item{from}{string (optional) The date (or date time combination) that data
is required from (if not provided only the last measurement will be
provided.)}

\item{to}{string (optional) The date (or date time combination) that data is
required up to (if not provided data will be provided from the from date to
the most recent data available.)}

\item{option}{string One of WQ, Standard or None.}

\item{timeInterval}{string (optional) A time interval that data is requested
over, either of the form P1D, or P3M, or fromDateTime/toDateTime.}

\item{alignment}{string (optional) A time or keyword that the data should be
aligned to.}

\item{method}{string (optional) A statistical method to apply to the data.
The options are Interpolate, Average, Total, Moving Average and EP.
Interpolate - Provide data aligned with the requested interval. Average -
Provide the average of the measurement across the interval. Total - Provide
the total in the interval. Moving Average - Provide the moving average in
the interval, timestamped to half of the interval. EP - Provide the
exceedence percentile based on the whole available dataset.}

\item{interval}{string (optional) The interval for the statistic to be
computed over.}
}
\value{
dataframe of measurements with optional metadata for requested sites.
}
\description{
\code{fullGetHilltopData} returns a dataframe of requested measurements at sites.
}
