% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_requests.R
\name{getHilltopData}
\alias{getHilltopData}
\title{General function to get timeseries data from a Hilltop server.}
\usage{
getHilltopData(endpoint, site, measurement, from, to, timeInterval, tsType,
  alignment, method, interval)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}

\item{site}{string A site that has data available on the server.}

\item{measurement}{string A measurement that has data available for the
requested site on the server.}

\item{from}{string (optional) The date (or date time combination) that data
is required from (if not provided only the last measurement will be
provided.)}

\item{to}{string (optional) The date (or date time combination) that data is
required up to (if not provided data will be provided from the from date to
the most recent data available.)}

\item{timeInterval}{string (optional) A time interval that data is requested
over, either of the form P1D, or P3M, or fromDateTime/toDateTime.}

\item{tsType}{string (optional) The type of time series being requested.
Default is StdTimeSeries, but quality code data can be obtained using
tsType = StdQualSeries.}

\item{alignment}{string (optional) A time or keyword that the data should be
aligned to.}

\item{method}{string (optional) A statistical method to apply to the data.
The options are Interpolate, Average, Total, Moving Average and EP.
Interpolate - Provide data aligned with the requested interval. Average -
Provide the average of the measurement across the interval. Total - Provide
the total in the interval. Moving Average - Provide the moving average in
the interval, timestamped to half of the interval. EP - Provide the
exceedence percentile based on the whole available dataset.}

\item{interval}{string (optional) The interval for the statistic to be
computed over.}
}
\value{
dataframe A dataframe of the timeseries data.
}
\description{
\code{getHilltopData} retrieves time series data from a Hilltop server.
}
\details{
Takes a valid Hilltop server endpoint, a site name, a measurement name  and
optional from, to or timeIntervaland arguments, an optional timeseries type
argument and returns a dataframe of the data for that site and measurement
for the requested time range from the endpoint. For standard time series,
statistics can be obtained, to request these a method, interval and
allignment argument need to be provided.
}
