% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_requests.R
\name{getHilltopSites}
\alias{getHilltopSites}
\title{General funtion to get a site list from a Hilltop server.}
\usage{
getHilltopSites(endpoint, measurement, location)
}
\arguments{
\item{endpoint}{string A valid hilltop endpoint.}

\item{measurement}{string (optional) a measurement name that a site must have available.}

\item{location}{logical (optional) whether the Lat Long for the sites is
returned}
}
\value{
A dataframe of site names
}
\description{
\code{getHilltopSites} returns a list of sites.
}
\details{
Takes a valid hilltop endpoint, and returns a list of available sites
available from the endpoint.
}
