% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{hilltopAttributeHelper}
\alias{hilltopAttributeHelper}
\title{Hilltop XML Attribute Identifier.}
\usage{
hilltopAttributeHelper(x)
}
\description{
Helper function to return the appropriate xml attribute name depending whether
the attribute of interest is from a named node, or is a named parameter.
}
