% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_measurements.R
\name{hilltopDataSourceToDF}
\alias{hilltopDataSourceToDF}
\title{Hilltop Data Source To Data Frame.}
\usage{
hilltopDataSourceToDF(xmldata)
}
\arguments{
\item{xmldata}{An XML document, returned from a url request using anyXmlParse.}
}
\value{
A dataframe of the Hilltop datasource information.
}
\description{
Helper function that reads the nodes within a the DataSource ItemInfo node of a
Hilltop XML response from a anyXmlParse(url) request such as
xmldata<-anyXmlParse(url). Returns a dataframe of the Info for each Item.
Handles missing results and doen't require prior knowledge of the items.
}
