% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_lists.R
\name{hilltopDsMeasList}
\alias{hilltopDsMeasList}
\title{Hilltop Data Source Information And Measurements At A Site.}
\usage{
hilltopDsMeasList(measlistxml)

hilltopDsMeasList(measlistxml)
}
\description{
Helper function.
Takes an xml document from a Hilltop MeasurementList at a Site request.
Returns a dataframe of the datasource information and measurements names.

Helper function.
Takes an xml document from a Hilltop MeasurementList at a Site request.
Returns a dataframe of the datasource information and measurements names.
}
