% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_lists.R
\name{hilltopDsMeasListFull}
\alias{hilltopDsMeasListFull}
\title{Hilltop Data Source Information And Measurement Information At A Site.}
\usage{
hilltopDsMeasListFull(measlistxml)

hilltopDsMeasListFull(measlistxml)
}
\description{
Takes an xml document from a Hilltop MeasurementList at a Site request.
Returns a dataframe of all of the datasource and measurement information
combined.

Takes an xml document from a Hilltop MeasurementList at a Site request.
Returns a dataframe of all of the datasource and measurement information
combined.
}
