% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_ensemble_stats.R
\name{hilltopEnsembleStatFull}
\alias{hilltopEnsembleStatFull}
\title{Hilltop Ensemble Statistics By Time Period, With Summary Information.}
\usage{
hilltopEnsembleStatFull(xmldata)
}
\arguments{
\item{xmldata}{An XML document, returned from a url request using anyXmlParse.}
}
\value{
A dataframe of the ensemble statistics along with the background metadata.
}
\description{
Takes the parsed xml from an EnsembleStats Request. Returns a dataframe of the
statistics for the period, along with the background information such as site
measurement units etc.
}
