% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_lists.R
\name{hilltopMeasInfoListExtra}
\alias{hilltopMeasInfoListExtra}
\title{Old Function To Get Data SOurce And Measurement Information.
Helper function.
Takes an xml document from a Hilltop MeasurementList at a Site request.
Returns a dataframe of the measurement information and datasource info.}
\usage{
hilltopMeasInfoListExtra(measlistxml)

hilltopMeasInfoListExtra(measlistxml)
}
\description{
Old Function To Get Data SOurce And Measurement Information.
Helper function.
Takes an xml document from a Hilltop MeasurementList at a Site request.
Returns a dataframe of the measurement information and datasource info.

Old Function To Get Data SOurce And Measurement Information.
Old Function To Get Data SOurce And Measurement Information.
Helper function.
Takes an xml document from a Hilltop MeasurementList at a Site request.
Returns a dataframe of the measurement information and datasource info.
}
