% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_measurements.R
\name{hilltopMeasurement}
\alias{hilltopMeasurement}
\title{Hilltop Measurement Request As A Data Frame.}
\usage{
hilltopMeasurement(dataxml)
}
\description{
Main function that converts a Hilltop XML document from a anyXmlParse(url)
request such as dataxml<-anyXmlParse(url) for a water quality measurement into
a dataframe that contains the measurement information. It returns a dataframe
of the data for each timestamp, including DataSource Information and the Site
Name. This dataframe can be merged with a WQ Sample dataframe processed using
hilltopMeasurementToDF
}
