% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_lists.R
\name{hilltopSiteList}
\alias{hilltopSiteList}
\title{Get A List of Sites Available From A Hilltop Web Service.}
\usage{
hilltopSiteList(xmldata)
}
\arguments{
\item{xmldata}{An XML document, returned from a url request using anyXmlParse.}
}
\value{
A dataframe of available sites (with locations if available).
}
\description{
Takes an parsed xml document from a Hilltop SiteList request.
Returns a dataframe of the available sites and location if available.
}
