% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{hilltopValueHelper}
\alias{hilltopValueHelper}
\title{Hilltop XML Value Identifier.}
\usage{
hilltopValueHelper(x)
}
\description{
Helper function to return the appropriate xml value depending whether the value
of interest is from a named node, or is a named parameter.
}
