% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{is.hilltopXml}
\alias{is.hilltopXml}
\title{Checks Whether XML Is Hilltop XML.}
\usage{
is.hilltopXml(xmldata)
}
\description{
Checks if an xml document is hilltop xml, returns True or False accordingly.
}
