% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{period}
\alias{period}
\title{Determine The TimePeriod Of Hilltop XML Ensemble Data.}
\usage{
period(dataxml)
}
\description{
Helper function to determine what the measurement period of the EnsembleStats is.
}
