% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillr_helpers.R
\name{stringEnd}
\alias{stringEnd}
\title{Return the last n charachters from a string.}
\usage{
stringEnd(string, n)
}
\arguments{
\item{string}{The string that the substring is to be extracted from.}

\item{n}{The number of charachters from the end of the string to return.}
}
\value{
A string that is the last n charachters from string.
}
\description{
Helper function to return the last n charachters of a string.
}
